requirements.txt
streamlit
google-generativeai
pypdf2
beautifulsoup4
requests
pymupdf